veripilot/pyproject.toml
[project]
name = "veripilot"
version = "0.1.0"
description = "Dual-Language Rust Verification Copilot for Lean 4 and Verus"
requires-python = ">=3.11"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "VeriPilot Team"}
]

dependencies = [
    # Lean Extraction
    "lean-dojo>=2.0.0",

    # RAG Framework
    "llama-index>=0.10.0",
    "llama-index-vector-stores-weaviate>=1.0.0",
    "llama-index-graph-stores-neo4j>=0.1.0",

    # Cloud Databases
    "weaviate-client>=4.0.0",
    "neo4j>=5.0.0",
    "duckdb>=0.10.0",

    # Embeddings - BGE (local, disk-conscious)
    "sentence-transformers>=2.2.0",

    # Utilities
    "pyyaml>=6.0",
    "pydantic>=2.0",
    "rich>=13.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.25.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.1.0",
    "mypy>=1.8",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
